Update README.md
American Sign Language (ASL) is the primary language used by many deaf individuals in North America, and it is also used by hard-of-hearing and hearing individuals. In this project I trained a convolutional neural network to classify images of ASL letters. After loading, examining, and preprocessing the data, trained the network and tested its performance.